Fixes with the BlankQuestPedestal
Fixed the error it was causing (I was tired and forgot to register the
pedestal)
Some other weird issues appear to be arising though, every time I
create a new blank pedestal, it seems to think there are 2 in one spot,
then 3, then 4 etc. Also when right clicking it with the quest book, it
seems to put a quest pedestal and a blank pedestal in the same spot, so
the quest pedestal doesn't work at all. I've gone through the other
pedestals for a solution but can't find anything.