Item Pedestals and fixes
Added a way to stop enderpearls and teleporting commands within the
skylands (Quick fix, still needs permissions for admins)
Added a way for the magic book to be added even if the inventory is
full.
Added an else for the item pedestal in the create pedestal command.
Syntax: /skyquest pedestal item <item id>
Also allows for damage values
Added get ItemStack to the pedestal base to get the floating item

(Forgot the saving of item pedestals for each player, will do that
later)